backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-chess==1.999
requests==2.31.0
pydantic==2.5.3
